SB-12812 # Pr comments resolved
