+++ Dockerfile
@@ -1,5 +1,4 @@
 FROM openjdk:8-jre-alpine
-MAINTAINER "Manojv" "manojv@ilimi.in"
 RUN apk update \
     && apk add  unzip \
     && apk add curl \
@@ -9,7 +8,7 @@ RUN apk update \
 #ENV sunbird_learnerstate_actor_port 8088 
 RUN chown -R sunbird:sunbird /home/sunbird
 USER sunbird
-COPY ./service/target/learning-service-1.0-SNAPSHOT-dist.zip /home/sunbird/learner/
-RUN unzip /home/sunbird/learner/learning-service-1.0-SNAPSHOT-dist.zip -d /home/sunbird/learner/
+COPY ./service/target/user-learning-service-1.0-SNAPSHOT-dist.zip /home/sunbird/learner/
+RUN unzip /home/sunbird/learner/user-learning-service-1.0-SNAPSHOT-dist.zip -d /home/sunbird/learner/
 WORKDIR /home/sunbird/learner/
-CMD java  -cp '/home/sunbird/learner/learning-service-1.0-SNAPSHOT/lib/*' play.core.server.ProdServerStart  /home/sunbird/learner/learning-service-1.0-SNAPSHOT
+CMD java  -cp '/home/sunbird/learner/user-learning-service-1.0-SNAPSHOT/lib/*' play.core.server.ProdServerStart  /home/sunbird/learner/user-learning-service-1.0-SNAPSHOT
