--- learning-service/.gitignore
@@ -1,6 +1,0 @@
-/target/
-/bin/
-/.classpath
-/.project
-/.settings
-/logs
