--- learning-service/conf/log4j2.properties
@@ -1,16 +1,0 @@
-status = info
-name = PropertiesConfig
-appenders = console
-
-appender.console.type = Console
-appender.console.name = STDOUT
-appender.console.layout.type = PatternLayout
-appender.console.layout.pattern = %m%n
-
-logger.default.name = defaultLogger
-logger.default.level = info
-
-
-rootLogger.level = debug
-rootLogger.appenderRefs = STDOUT
-rootLogger.appenderRef.stdout.ref = STDOUT
