--- learning-service/conf/routes
@@ -1,29 +1,0 @@
-# Routes
-# This file defines all application routes (Higher priority routes first)
-# ~~~~
-
-# User management APIs
-POST	/v1/user/create				            @controllers.usermanagement.UserController.createUser()
-POST	/v2/user/create				            @controllers.usermanagement.UserController.createUserV2()
-PATCH	/v1/user/update					          @controllers.usermanagement.UserController.updateUser()
-PATCH	/private/user/v1/update					          @controllers.usermanagement.UserController.updateUser()
-GET  	/v1/user/read/:uid				        @controllers.usermanagement.UserController.getUserById(uid:String)
-GET  	/v2/user/read/:uid				        @controllers.usermanagement.UserController.getUserByIdV2(uid:String)
-POST  /v1/user/getuser                  @controllers.usermanagement.UserController.getUserByLoginId()
-GET   /v1/user/get/:idType/:id          @controllers.usermanagement.UserController.getUserByKey(idType:String,id:String)
-POST  /v1/user/search					 @controllers.usermanagement.UserController.searchUser()
-POST  /private/user/v1/search            @controllers.usermanagement.UserController.searchUser()
-GET   /v1/user/mediatype/list     @controllers.usermanagement.UserProfileController.getProfileSupportedSocialMediaTypes()
-POST  /v1/user/profile/visibility @controllers.usermanagement.UserProfileController.setProfileVisibility()
-
-POST  /v1/user/block			        @controllers.usermanagement.UserStatusController.blockUser()
-POST  /v1/user/unblock			      @controllers.usermanagement.UserStatusController.unblockUser()
-
-GET   /v1/role/read						    @controllers.usermanagement.UserRoleController.getRoles()
-POST  /v1/user/assign/role			  @controllers.usermanagement.UserRoleController.assignRoles()
-POST  /private/user/v1/assign/role	  @controllers.usermanagement.UserRoleController.assignRoles()
-
-PATCH /v1/user/update/logintime	  @controllers.usermanagement.UserLoginController.updateLoginTime()
-
-POST  /v1/user/tnc/accept         @controllers.tac.UserTnCController.acceptTnC()
-GET   /v1/user/type/list		  @controllers.usermanagement.UserTypeController.getUserTypes()
