+++ metadata.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 # return version
-echo '{"name":"learner_service","version":"1.15.0","org":"sunbird","hubuser":"purplesunbird"}'
+echo '{"name":"user_learner_service","version":"1.15.0","org":"sunbird","hubuser":"purplesunbird"}'
