+++ pom.xml
@@ -4,7 +4,7 @@
   <artifactId>lms-service</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>pom</packaging>
-  <name>sunbird_lms_service</name>
+  <name>user-learning-service</name>
 
   <properties>
 	  <maven.compiler.source>1.8</maven.compiler.source>
@@ -13,6 +13,6 @@
 
   <modules>
 	  <!--<module>actors/sunbird-lms-mw</module>-->
-	  <module>learning-service</module>
+	  <module>user-learning-service</module>
   </modules>
 </project>
